For this project, I've created an object-oriented Java application for managing student grade reports. The program incorporates encapsulation, inheritance, composition, and polymorphism to enhance functionality and maintainability. Encapsulation is used to encapsulate data and behavior within classes. Inheritance is seen in the relationship between the "Student" and "Person" classes. Composition is employed in the "GradeReport" class, which contains a list of "SubjectScore" objects. Polymorphism is demonstrated in the "SubjectScore" class, allowing for different grading criteria. Overall, these principles contribute to an efficient and adaptable system for managing student grades.
For this project, I've created an object-oriented Java application for managing student grade reports. The program incorporates encapsulation, inheritance, composition, and polymorphism to enhance functionality and maintainability. Encapsulation is used to encapsulate data and behavior within classes. Inheritance is seen in the relationship between the "Student" and "Person" classes. Composition is employed in the "GradeReport" class, which contains a list of "SubjectScore" objects. Polymorphism is demonstrated in the "SubjectScore" class, allowing for different grading criteria. Overall, these principles contribute to an efficient and adaptable system for managing student grades.